Add files via upload
added a top of page button to the about page. pretty sure I covered all the bases of what was asked for for the project in the last commit but wanted to add a top button. It didnt come out the way I wanted but i eventually got it to work.